Been a while since committing...
Got too worried about optimization, but re-aligned goals to reflect that of an MVP.

Currently the Search has to be finished being updated with the new implementation.

App now clears cache on start up. Still need to manage the size of the cache while that app is running, and handling live updates... Currently the app will only show live changes if it is restarted (which may be best, or i'll add a refresh tool [swipe up] to clear the cache and refetch data.)