Implemented a suggested events POST bloc to communicate with the database repository provider to grab records from firebase using a paginated query.
Also refactored the ClubBigCard to work with the home bloc as well...

TODO: Implement a scroll controller on the home screen to automatically grab more records when scrolling to the end of the sliver list (maybe look up SliverList controllers.)

TODO: Have the Home Bloc and Account Blocs maintain state between the two. So when switching from home to account the blocs won't get closed..

TODO: Refactor the eventDetails to its own screen and implement Slivers for the image and event.