Suggested Events now shows upcoming events (I plan on switching to firebase cloud functions to search for events occuring within the next few days, then withing a week, then within a month, etc.
Accounts Page now shows events that account created.
Accounts Page shows an image upload progress when the event is submitted.

"Posting" a new event now automatically redirects you to the account page, to show the upload progress.
Uploading an image, now shows a message when access is denied to the photo gallery. Prompts user to go to settings.
When app is resumed from a paused state, the upload image section will try to automatically retrieve the device's images again.
Failed retrieval of device images now offers a "retry" button, for the user to send another DeviceImageFetch event.
Added a loading widget when fetching events is in progress.

Updated the EventModel, SearchResultModel, and the way BloC's interact with them, to follow dart conventions.
Simplified the database repository by adding constants for the db attributes, and abstracting the db functions.

The home and account page's are now implemented as a page view in order to allow for their states to be remembered when "navigating" (technically, "tabbing", between them.

Added cute messages to the home and accounts page, when the suggested events bloc and account events bloc retrieve empty results.

And much, much, more!