Users can now update events.
When the user clicks on the "More Vert" Icon Button on the Account Screen, that event triggers fetching the entire event which should finish by the time the modal bottom sheet appears.

Also Added more documentation and made the Database and authentication repository follow singleton rules

TODO: What happens if it fails to retrieve the full event?